chore: rename AuthViewModel to AuthenticatorState
