+++ packages/amplify_authenticator/lib/src/enums/authenticator_step.dart
@@ -14,6 +14,8 @@
  */
 
 enum AuthenticatorStep {
+  loading,
+  authenticated,
   initial,
   signup,
   signin,
