+++ packages/amplify_authenticator/lib/src/l10n/title_resolver.dart
@@ -85,6 +85,9 @@ class TitleResolver extends Resolver<AuthenticatorStep> {
       case AuthenticatorStep.verifyUser:
       case AuthenticatorStep.confirmVerifyUser:
         return verifyUser(context);
+      case AuthenticatorStep.loading:
+      case AuthenticatorStep.authenticated:
+        throw StateError('Invalid step: $this');
     }
   }
 }
